+++ routes/web.php
@@ -3,8 +3,6 @@
 use App\Http\Controllers\DocumentController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::get('/', [DocumentController::class, 'create'])->name('documents.create');
 
 Route::post('/documents', [DocumentController::class, 'store'])->name('documents.store');